Refactor rayon-core into focused modules
