+++ crates/core/src/launcher/mod.rs
@@ -1,0 +1,12 @@
+mod error;
+mod execution;
+mod search;
+mod service;
+mod sessions;
+mod state;
+
+#[cfg(test)]
+mod tests;
+
+pub use error::LauncherError;
+pub use service::LauncherService;
